auto progress
the tick_per_inch value was (when rounded) 48.8. however, there was a lot of inconsistency with the value, so i changed it to 26.25, a more consistent number, for the time being, commenting out the original value. keep in mind this is a tentative number
also the robot can pick up cones autonomously now